feat: add round counting system that is displayed Add property roundNumber to object gameState.
Add logic to function startGame to display round number in the console
when game starts.

Add logic to increment roundNumber by 1 in function nextRound, and
display round number in console when the next round is entered.

Add logic to function handleNextRoundClick to display the round number
in the game display when game starts and when a new round is entered.