feat: add game over logic when a player score = 3 Add logic to end the game when a player reaches the score of 3 points first.
Add property isGameOver to gameState.

Add function checkGameOver to module gameControllerTTT to check if a
player has 3 points. If true, it displays (in the console) that the game
is over and declares the game winner.

Call function checkGameOver in function playRound after every instance
of a player winning a round to check if win condition of 3 points has
been met.